# src/main/resources/openapi/customer-api.yaml
openapi: 3.0.3

info:
  title: Customer API
  version: v1

tags:
  - name: Customer
    description: Customer operations

paths:
  /api/v1/customers:
    get:
      operationId: findAllCustomers
      tags:
        - Customer
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
        - name: type
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/CustomerTypeRequest'
      responses:
        '200':
          description: Customers found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerPageResponse'
    post:
      operationId: createCustomer
      tags:
        - Customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomerRequest'
      responses:
        '201':
          description: Customer created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'

  /api/v1/customers/{customerId}:

    parameters:
      - name: customerId
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: findCustomerById
      tags:
        - Customer
      responses:
        '200':
          description: Customer found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '400':
          description: Customer not found - the customerId does not exist

    put:
      operationId: updateCustomer
      tags:
        - Customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCustomerRequest'
      responses:
        '200':
          description: Customer updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '400':
          description: Invalid request - customer is inactive or invalid fields provided
        '404':
          description: Customer not found

    delete:
      operationId: deleteCustomer
      tags:
        - Customer
      responses:
        '204':
          description: Customer deleted

components:
  schemas:
    CreateCustomerRequest:
      type: object
      required:
        - customerType
        - documentType
        - documentNumber
      properties:
        customerType:
          $ref: '#/components/schemas/CustomerTypeRequest'
        documentType:
          $ref: '#/components/schemas/DocumentType'
        documentNumber:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        businessName:
          type: string
        email:
          type: string
          format: email
        phoneNumber:
          type: string

    CustomerResponse:
      type: object
      properties:
        customerId:
          type: string
        customerType:
          $ref: '#/components/schemas/CustomerTypeRequest'
        documentType:
          $ref: '#/components/schemas/DocumentType'
        documentNumber:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        businessName:
          type: string
        email:
          type: string
        phoneNumber:
          type: string

    UpdateCustomerRequest:
      type: object
      properties:
        documentNumber:
          type: string
          description: Only updatable for PERSONAL customers
        firstName:
          type: string
          description: Only updatable for PERSONAL customers
        lastName:
          type: string
          description: Only updatable for PERSONAL customers
        businessName:
          type: string
          description: Not updatable after creation
        email:
          type: string
          format: email
          description: Updatable for both PERSONAL and BUSINESS customers
        phoneNumber:
          type: string
          description: Updatable for both PERSONAL and BUSINESS customers

    CustomerTypeRequest:
      type: string
      enum:
        - PERSONAL
        - BUSINESS

    DocumentType:
      type: string
      enum:
        - DNI
        - CE
        - PASSPORT
        - RUC

    CustomerPageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/CustomerResponse'
        pageNumber:
          type: integer
        pageSize:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
        isLast:
          type: boolean